refactor(env): break component to files
