+++ app/react/portainer/environments/ItemView/UpdateForm/AzureForm/AzureConfiguration.tsx
