+++ app/react/portainer/environments/ItemView/UpdateForm/EdgeForm/AMTInfo.tsx
@@ -4,8 +4,8 @@ import { FormControl } from '@@/form-components/FormControl';
 import { FormSection } from '@@/form-components/FormSection';
 import { Input } from '@@/form-components/Input';
 
-import { EnvironmentId } from '../../types';
-import { useSettings } from '../../../settings/queries';
+import { EnvironmentId } from '../../../types';
+import { useSettings } from '../../../../settings/queries';
 
 export function AmtInfo({ environmentId }: { environmentId: EnvironmentId }) {
   const isAmtEnabledQuery = useSettings(
